requirements.txt
google-generativeai
python-docx
PyMuPDF
fpdf
python-dotenv